refactor: ♻️ Updating Infoblox integration for Nautobot 2.0 changes.
As part of the changes removed Aggregate objects as that is no longer a valid model in 2.0. Updated all Site creation to be a Location. Also removed the associated setting for importing Aggregates.

Updates for Nautobot 2.0 so lots of breaking changes. No longer an Aggregate object to import and removed extraneous Job. The `import_rfc1918" setting is no longer valid.