Move cache to adapter and validate response
This change moves the cache to the adapter, which means there will be a new cache every job run, instead of using a cache from when the server starts.
Also adds checks and debug messages for the create/update so we can see when there are issues creating or updating.